+++ src/client/java/net/detectivekaktus/client/data/DotcItemTagProvider.java
@@ -50,7 +50,10 @@ public class DotcItemTagProvider extends FabricTagProvider<Item> {
                 .add(DotcTools.BLINK_DAGGER)
                 .add(DotcTools.SWIFT_BLINK)
                 .add(DotcTools.ARCANE_BLINK)
-                .add(DotcTools.OVERWHELMING_BLINK);
+                .add(DotcTools.OVERWHELMING_BLINK)
+                .add(DotcPrimitives.OBLIVION_STAFF)
+                .add(DotcTools.ORCHID_MALEVOLENCE)
+                .add(DotcTools.BLOODTHORN);
 
         getOrCreateTagBuilder(ItemTags.AXES)
                 .add(DotcPrimitives.OGRE_AXE)