+++ src/client/java/net/detectivekaktus/client/data/item/tool/DotcToolsModelProvider.java
@@ -42,7 +42,9 @@ public class DotcToolsModelProvider extends FabricModelProvider {
                 DotcTools.BLINK_DAGGER,
                 DotcTools.SWIFT_BLINK,
                 DotcTools.ARCANE_BLINK,
-                DotcTools.OVERWHELMING_BLINK
+                DotcTools.OVERWHELMING_BLINK,
+                DotcTools.ORCHID_MALEVOLENCE,
+                DotcTools.BLOODTHORN
         ).forEach(item -> itemModelGenerator.generateFlatItem(item, ModelTemplates.FLAT_HANDHELD_ITEM));
 
         Stream.of(