+++ src/main/java/net/detectivekaktus/core/item/ItemStackHelper.java
@@ -1,11 +1,7 @@
 package net.detectivekaktus.core.item;
 
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResult;
-import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 
 import net.detectivekaktus.attach.PlayerFlags;
@@ -14,24 +10,30 @@ import net.detectivekaktus.effect.DotcEffects;
 import net.detectivekaktus.sound.gui.DotcGuiSounds;
 
 public class ItemStackHelper {
-    public static InteractionResultHolder<ItemStack> cancelInteractionIfStunned(Player player, Level level, InteractionHand hand) {
-        if (!player.hasEffect(DotcEffects.STUN))
-            return null;
-
-        if (level.isClientSide)
-            player.playSound(DotcGuiSounds.UI_GENERAL_DENY);
-
-        return InteractionResultHolder.fail(player.getItemInHand(hand));
+    public static boolean cancelInteractionIfDisabled(Player player) {
+        if (cancelInteractionIfStunned(player))
+            return true;
+        return cancelInteractionIfSilenced(player);
     }
 
-    public static InteractionResult cancelInteractionIfStunned(Player player) {
+    private static boolean cancelInteractionIfStunned(Player player) {
         if (!player.hasEffect(DotcEffects.STUN))
-            return null;
+            return false;
 
         if (player.level().isClientSide)
             player.playSound(DotcGuiSounds.UI_GENERAL_DENY);
 
-        return InteractionResult.FAIL;
+        return true;
+    }
+
+    private static boolean cancelInteractionIfSilenced(Player player) {
+        if (!player.hasEffect(DotcEffects.SILENCE) && !player.hasEffect(DotcEffects.SOUL_REND))
+            return false;
+
+        if (player.level().isClientSide)
+            player.playSound(DotcGuiSounds.UI_SILENCE);
+
+        return true;
     }
 
     public static ShadowWalkingSource revealInvisibility(Player player) {