+++ src/main/java/net/detectivekaktus/effect/DotcEffects.java
@@ -13,6 +13,8 @@ public class DotcEffects {
     public static final Holder<MobEffect> DISARM = register("disarm", new Disarm());
     public static final Holder<MobEffect> BREAK = register("break", new Break());
     public static final Holder<MobEffect> COMBO_BREAKER = register("combo_breaker", new ComboBreaker());
+    public static final Holder<MobEffect> SILENCE = register("silence", new Silence());
+    public static final Holder<MobEffect> SOUL_REND = register("soul_rend", new Silence());
 
     public static Holder<MobEffect> register(String id, MobEffect effect) {
         return Registry.registerForHolder(