+++ src/main/java/net/detectivekaktus/sound/item/DotcItemSounds.java
@@ -25,6 +25,9 @@ public class DotcItemSounds {
     public static final SoundEvent ARCANE_BLINK_TARGET = DotcSounds.register("item_arcane_blink_target");
     public static final SoundEvent OVERWHELMING_BLINK_TARGET = DotcSounds.register("item_overwhelming_blink_target");
     public static final SoundEvent AEON_DISK = DotcSounds.register("item_aeon_disk");
+    public static final SoundEvent ORCHID_MALEVOLENCE = DotcSounds.register("item_orchid_malevolence");
+    public static final SoundEvent BLOODTHORN = DotcSounds.register("item_bloodthorn");
+    public static final SoundEvent BLOODTHORN_TARGET = DotcSounds.register("item_bloodthorn_target");
 
     public static void initialize() {}
 }